Add run scripts for building ballerina project

+++ scripts/ballerina-pack/build.sh
@@ -1,7 +1,11 @@
 #!/usr/bin/env bash
 
-source "../properties.sh"
-source "../utils.sh"
+DEV_BALLERINA_CURRENT_SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
+
+# shellcheck source=../properties.sh
+source "${DEV_BALLERINA_CURRENT_SCRIPT_DIR}/../properties.sh"
+# shellcheck source=../utils.sh
+source "${DEV_BALLERINA_CURRENT_SCRIPT_DIR}/../utils.sh"
 
 echo "Building Ballerina Repo"
 runBallerinaLangGradleBuild clean build \
@@ -11,14 +15,18 @@ runBallerinaLangGradleBuild clean build \
   "$@"
 
 DEV_BALLERINA_PACK_ZIP=${DEV_BALLERINA_PACK}.zip
-echo "Removing previous Ballerina Pack zip ${DEV_BALLERINA_PACK_ZIP}"
-rm -f "${DEV_BALLERINA_PACK_ZIP}"
+if [ -f "${DEV_BALLERINA_PACK_ZIP}" ]; then
+  echo "Removing previous Ballerina Pack zip ${DEV_BALLERINA_PACK_ZIP}"
+  rm -f "${DEV_BALLERINA_PACK_ZIP}"
+fi
 
 echo "Copying new Ballerina Pack zip to ${DEV_BALLERINA_PACK_ZIP}"
 cp  "${DEV_BALLERINA_REPO}/distribution/zip/jballerina-tools/build/distributions/${DEV_BALLERINA_PACK_NAME}.zip" "${DEV_BALLERINA_PACK_ZIP}"
 
-echo "Removing previous Ballerina Pack ${DEV_BALLERINA_PACK}"
-rm -rf "${DEV_BALLERINA_PACK}"
+if [ -d "${DEV_BALLERINA_PACK}" ]; then
+  echo "Removing previous Ballerina Pack ${DEV_BALLERINA_PACK}"
+  rm -rf "${DEV_BALLERINA_PACK}"
+fi
 
 BALLERINA_PACK_DIR=$(dirname "${DEV_BALLERINA_PACK}")
 echo "Unzipping new Ballerina Pack to ${BALLERINA_PACK_DIR}/${DEV_BALLERINA_PACK_NAME}"
